requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
pillow
drf_yasg
